Fix for scheduled minion jobs that hangs forever
3005 introduced a lot of improvements to the IPC channels, unfortunately
it also introduced a bug where minion events are sent twice to the
master. In some cases where the event payload is bigger than 8k bytes,
which in our case happens everytime we do state.apply as a scheduled
event, the minion will get stuck in its send loop when its sending the
second event.

The send loop waits for another process to read the unix_socket but I
belive that process got reaped when the event got sent the first time,
so it will be stuck in its send loop forever.

However if we successfully write all data to the socket on the first
iteration when we are sending the event the second time the loop will end
and this in turn triggers a new process to start and read the socket.

Our state.apply messages are very big so this commit increases the
socket send buffer to 2^19 from 8k.

CES-546